Minor Spelling/Grammar Corrections
***NOTE: My version only shows the styled layout on the default page,
but none to the inner pages are rendered in that clipboard style layout.
However, I see the on-line version does behave as such.  Is something
out of sync? Or is it my env?

Also on some, it looks like a bigger changes, but that is just because
my IDE cleansed some of the intermix of tabs & spaces in the code
layout.